expose Core::new_with_base as reusable code.
